Renamed compress2 to compression
